backend/requirements.txt
fastapi
uvicorn
websockets
psycopg2-binary
sqlalchemy
redis
pydantic
python-dotenv
google-genai
python-multipart
